move modules up out of api module and add new clients module, break up orders client
