+++ setup.py
@@ -15,7 +15,7 @@ from pathlib import Path
 from setuptools import setup, find_packages
 
 
-with open('planet/api/__version__.py') as f:
+with open('planet/__version__.py') as f:
     for line in f:
         if line.find("__version__") >= 0:
             version = line.split("=")[1].strip()
